en principio se termino, ahora a mejorar todo
