+++ CSV/ratings.csv
@@ -3,52 +3,52 @@ Black Panther,0.0
 Avengers: Endgame,0.0
 Mission: Impossible -- Fallout,0.0
 Mad Max: Fury Road,0.0
-Spider-Man: Into the Spider-Verse,7.0
+Spider-Man: Into the Spider-Verse,9.0
 Wonder Woman,0.0
 Dunkirk,0.0
-Coco,0.0
-Thor: Ragnarok,6.0
+Coco,9.0
+Thor: Ragnarok,0.0
 Logan,0.0
 Star Wars: The Last Jedi,0.0
 Star Wars: The Force Awakens,0.0
-The Adventures of Robin Hood,0.0
+The Adventures of Robin Hood,8.0
 King Kong,0.0
-Spider-Man: Far From Home,10.0
+Spider-Man: Far From Home,0.0
 Incredibles 2,0.0
-Zootopia,0.0
+Zootopia,10.0
 War for the Planet of the Apes,0.0
-Spider-Man: Homecoming,10.0
+Spider-Man: Homecoming,0.0
 Baby Driver,0.0
 Seven Samurai,0.0
 Metropolis,0.0
 Jaws,0.0
-Up,10.0
+Up,9.0
 Shazam!,0.0
-The Dark Knight,8.0
+The Dark Knight,0.0
 The Treasure of the Sierra Madre,0.0
 Blade Runner 2049,0.0
 The 39 Steps,0.0
 Captain America: Civil War,0.0
 The Jungle Book,0.0
-Harry Potter and the Deathly Hallows: Part 2,10.0
-Skyfall,10.0
+Harry Potter and the Deathly Hallows: Part 2,0.0
+Skyfall,0.0
 Lawrence of Arabia,0.0
-Mission: Impossible Rogue Nation,3.0
+Mission: Impossible Rogue Nation,0.0
 Apocalypse Now,0.0
 The Hurt Locker,0.0
-The LEGO Movie,0.0
+The LEGO Movie,10.0
 Star Trek,0.0
 Moana,0.0
 Avengers: Infinity War,0.0
-Ant-Man and The Wasp,8.0
+Ant-Man and The Wasp,0.0
 Star Wars,0.0
 Isle of Dogs,0.0
 Stagecoach,0.0
 Goldfinger,0.0
-Guardians of the Galaxy,9.0
+Guardians of the Galaxy,0.0
 Kubo and the Two Strings,0.0
 WALL-E,0.0
-Marvel's the Avengers,6.0
+Marvel's the Avengers,0.0
 Iron Man,0.0
 The Terminator,0.0
 The African Queen,0.0
@@ -56,8 +56,8 @@ Rio Bravo,0.0
 Hunt for the Wilderpeople,0.0
 The French Connection,0.0
 The Taking of Pelham One Two Three,0.0
-John Wick: Chapter 3 -- Parabellum,10.0
-Doctor Strange,10.0
+John Wick: Chapter 3 -- Parabellum,0.0
+Doctor Strange,0.0
 Aliens,0.0
 The Empire Strikes Back,0.0
 Paths of Glory,0.0
@@ -73,12 +73,12 @@ Looper,0.0
 Chicken Run,0.0
 Stalker,0.0
 Casino Royale,0.0
-Guardians of the Galaxy Vol. 2,10.0
-The Peanut Butter Falcon,3.0
+Guardians of the Galaxy Vol. 2,0.0
+The Peanut Butter Falcon,0.0
 No Country for Old Men,0.0
 Dawn of the Planet of the Apes,0.0
 Captain Marvel,0.0
-Onward,0.0
+Onward,8.0
 Sicario,0.0
 Ghostbusters,0.0
 Ran,0.0
@@ -92,7 +92,7 @@ The Searchers,0.0
 Deadpool 2,0.0
 Aguirre: The Wrath of God,0.0
 "Crouching Tiger, Hidden Dragon",0.0
-The Fugitive,7.0
+The Fugitive,0.0
 Apollo 13,0.0
 The Lion King,0.0
 Bullitt,0.0
@@ -104,11 +104,11 @@ Inside Out,10.0
 Snow White and the Seven Dwarfs,0.0
 Paddington 2,0.0
 Toy Story 3,0.0
-Toy Story 2,0.0
+Toy Story 2,9.0
 Pinocchio,0.0
-Finding Dory,0.0
+Finding Dory,9.0
 Finding Nemo,9.0
-Toy Story,0.0
+Toy Story,8.0
 How to Train Your Dragon,10.0
 Shaun the Sheep Movie,0.0
 Dumbo,0.0
@@ -120,7 +120,7 @@ One Hundred and One Dalmatians,0.0
 Beauty and the Beast,0.0
 The Nightmare Before Christmas,0.0
 Anomalisa,0.0
-"Monsters, Inc.",9.0
+"Monsters, Inc.",10.0
 Grave of the Fireflies,0.0
 Fantasia,0.0
 Your Name,0.0
@@ -129,7 +129,7 @@ The Iron Giant,0.0
 Song of the Sea,0.0
 Tower,0.0
 Waltz With Bashir,0.0
-How to Train Your Dragon: The Hidden World,10.0
+How to Train Your Dragon: The Hidden World,9.0
 Persepolis,0.0
 Wallace & Gromit: The Curse of the Were-Rabbit,0.0
 Yellow Submarine,0.0
@@ -139,10 +139,10 @@ Ernest & Célestine,0.0
 Aladdin,0.0
 The Secret World of Arrietty,0.0
 Frozen,0.0
-Ralph Breaks the Internet,0.0
+Ralph Breaks the Internet,9.0
 The Red Turtle,0.0
 Long Way North,0.0
-How to Train Your Dragon 2,0.0
+How to Train Your Dragon 2,10.0
 Coraline,0.0
 Ghost in the Shell,0.0
 The Breadwinner,0.0
@@ -157,21 +157,21 @@ Princess Mononoke,0.0
 Bambi,0.0
 Arthur Christmas,0.0
 Antz,0.0
-The LEGO Movie 2: The Second Part,0.0
-A Bug's Life,0.0
+The LEGO Movie 2: The Second Part,10.0
+A Bug's Life,7.0
 Mary and Max,0.0
 Ponyo,0.0
-Klaus,6.0
-Teen Titans GO! to the Movies,8.0
+Klaus,0.0
+Teen Titans GO! to the Movies,0.0
 Bolt,0.0
 Tangled,0.0
-Shrek 2,0.0
+Shrek 2,9.0
 ParaNorman,0.0
 Frankenweenie,0.0
 James and the Giant Peach,0.0
 Rango,0.0
 Shrek,0.0
-Missing Link,1.0
+Missing Link,0.0
 The Wind Rises,0.0
 Winnie the Pooh,0.0
 The Simpsons Movie,0.0
@@ -179,10 +179,10 @@ Boy and the World,0.0
 The Peanuts Movie,0.0
 Millennium Actress,0.0
 When Marnie Was There,0.0
-Kung Fu Panda,10.0
-Sleeping Beauty,6.0
-Wreck-It Ralph,0.0
-Kung Fu Panda 3,10.0
+Kung Fu Panda,0.0
+Sleeping Beauty,0.0
+Wreck-It Ralph,10.0
+Kung Fu Panda 3,0.0
 The Illusionist,0.0
 Parasite,0.0
 The Cabinet of Dr. Caligari,0.0
@@ -343,7 +343,7 @@ Marty,0.0
 One Flew Over the Cuckoo's Nest,0.0
 Mary Poppins,0.0
 Some Like It Hot,0.0
-Lady Bird,7.0
+Lady Bird,0.0
 BlacKkKlansman,0.0
 Eighth Grade,0.0
 Booksmart,0.0
@@ -382,7 +382,7 @@ To Be or Not to Be,0.0
 Groundhog Day,0.0
 The Player,0.0
 You Can't Take It With You,0.0
-The Truman Show,7.0
+The Truman Show,0.0
 Big,0.0
 Enough Said,0.0
 The Apartment,0.0
@@ -422,7 +422,7 @@ Jodorowsky's Dune,0.0
 Welcome to Chechnya,0.0
 Shirkers,0.0
 Anvil! The Story of Anvil,0.0
-Inside Job,6.0
+Inside Job,0.0
 All In: The Fight for Democracy,0.0
 The Square,0.0
 Seymour: An Introduction,0.0
@@ -544,7 +544,7 @@ Possessor: Uncut,0.0
 Relic,0.0
 The Birds,0.0
 Young Frankenstein,0.0
-Ready or Not,10.0
+Ready or Not,0.0
 La llorona,0.0
 Drag Me to Hell,0.0
 A Girl Walks Home Alone at Night,0.0
@@ -609,18 +609,18 @@ My Fair Lady,0.0
 Hugo,0.0
 Enchanted,0.0
 The Straight Story,0.0
-Harry Potter and the Prisoner of Azkaban,10.0
+Harry Potter and the Prisoner of Azkaban,0.0
 Pete's Dragon,0.0
 The Secret of Roan Inish,0.0
 A Monster Calls,0.0
 Spy Kids,0.0
 Summer Hours,0.0
-Harry Potter and the Goblet of Fire,9.0
+Harry Potter and the Goblet of Fire,0.0
 The Kid Who Would Be King,0.0
 How Green Was My Valley,0.0
 That Thing You Do!,0.0
 First Position,0.0
-Willy Wonka and the Chocolate Factory,0.0
+Willy Wonka and the Chocolate Factory,8.0
 Lassie,0.0
 Duma,0.0
 Jason and the Argonauts,0.0
@@ -907,7 +907,7 @@ Phil Spector,0.0
 I Am,0.0
 The Real Cancun,0.0
 15 Minutes,0.0
-Man of the Year,7.0
+Man of the Year,0.0
 Pokémon 3: The Movie,0.0
 The Honeymooners,0.0
 From Justin to Kelly,0.0
@@ -975,7 +975,7 @@ The Lone Ranger,0.0
 The Alamo,0.0
 The Warrior's Way,0.0
 The Dark Tower,0.0
-Wild Wild West,5.0
+Wild Wild West,0.0
 Priest,0.0
 September Dawn,0.0
 American Outlaws,0.0
